modified:   .vscode/settings.json 	modified:   Practice/modular_inverse/modular_inverse/src/continued_fraction.cpp 	modified:   Practice/modular_inverse/modular_inverse/src/euclid.cpp 	modified:   Practice/modular_inverse/modular_inverse/src/fermat.cpp 	modified:   Practice/modular_inverse/modular_inverse/src/fermat_exponentiaon.cpp 	modified:   Practice/modular_inverse/modular_inverse/src/menu.cpp 	modified:   Practice/modular_inverse/modular_inverse/src/modular_inverse.cpp 	modified:   Practice/modular_inverse/modular_inverse/src/rsa.cpp 	modified:   digit_ops/digit_ops/obj/Debug/net9.0/digit_ops.AssemblyInfo.cs 	modified:   digit_ops/digit_ops/obj/Debug/net9.0/digit_ops.AssemblyInfoInputs.cache 	modified:   digit_ops/digit_ops/obj/Debug/net9.0/digit_ops.GeneratedMSBuildEditorConfig.editorconfig 	modified:   digit_ops/digit_ops/obj/Debug/net9.0/digit_ops.assets.cache 	modified:   digit_ops/digit_ops/obj/digit_ops.csproj.nuget.dgspec.json 	modified:   digit_ops/digit_ops/obj/digit_ops.csproj.nuget.g.props 	modified:   digit_ops/digit_ops/obj/project.assets.json 	modified:   digit_ops/digit_ops/obj/project.nuget.cache 	modified:   pangram/pangram/obj/Debug/net9.0/pangram.AssemblyInfo.cs 	modified:   pangram/pangram/obj/Debug/net9.0/pangram.AssemblyInfoInputs.cache 	modified:   pangram/pangram/obj/Debug/net9.0/pangram.GeneratedMSBuildEditorConfig.editorconfig 	modified:   pangram/pangram/obj/Debug/net9.0/pangram.assets.cache 	modified:   pangram/pangram/obj/pangram.csproj.nuget.dgspec.json 	modified:   pangram/pangram/obj/pangram.csproj.nuget.g.props 	modified:   pangram/pangram/obj/project.assets.json 	modified:   pangram/pangram/obj/project.nuget.cache 	modified:   unvisited_stones/unvisited_stones/obj/Debug/net9.0/unvisited_stones.AssemblyInfo.cs 	modified:   unvisited_stones/unvisited_stones/obj/Debug/net9.0/unvisited_stones.AssemblyInfoInputs.cache 	modified:   unvisited_stones/unvisited_stones/obj/Debug/net9.0/unvisited_stones.GeneratedMSBuildEditorConfig.editorconfig 	modified:   unvisited_stones/unvisited_stones/obj/Debug/net9.0/unvisited_stones.assets.cache 	modified:   unvisited_stones/unvisited_stones/obj/project.assets.json 	modified:   unvisited_stones/unvisited_stones/obj/project.nuget.cache 	modified:   unvisited_stones/unvisited_stones/obj/unvisited_stones.csproj.nuget.dgspec.json 	modified:   unvisited_stones/unvisited_stones/obj/unvisited_stones.csproj.nuget.g.props

+++ Practice/modular_inverse/modular_inverse/src/fermat.cpp
@@ -1,7 +1,13 @@
 #include "fermat.h"
 #include <cmath>
 #include <stdexcept>
+#include <cstdint> // Для int64_t
 
+/**
+ * Проверяет, является ли число простым.
+ * @param n Число для проверки.
+ * @return true, если число простое, иначе false.
+ */
 bool is_prime(int n) {
     if (n <= 1) return false;
     if (n == 2) return true;
@@ -16,10 +22,17 @@ bool is_prime(int n) {
     return true;
 }
 
-long long mod_exp(long long base, long long exponent, long long mod) {
+/**
+ * Вычисляет base^exponent mod mod.
+ * @param base Основание.
+ * @param exponent Показатель степени.
+ * @param mod Модуль.
+ * @return Результат возведения в степень по модулю.
+ */
+int64_t mod_exp(int64_t base, int64_t exponent, int64_t mod) {
     if (mod == 1) return 0;
     
-    long long result = 1;
+    int64_t result = 1;
     base = base % mod;
     
     while (exponent > 0) {
@@ -34,7 +47,14 @@ long long mod_exp(long long base, long long exponent, long long mod) {
     return result;
 }
 
-long long modInverseFermat(long long a, long long p) {
+/**
+ * Находит обратный элемент по модулю с использованием теоремы Ферма.
+ * @param a Число, для которого ищется обратный элемент.
+ * @param p Простой модуль.
+ * @return Обратный элемент.
+ * @throws invalid_argument Если модуль не является простым числом или число кратно модулю.
+ */
+int64_t modInverseFermat(int64_t a, int64_t p) {
     if (!is_prime(p)) {
         throw std::invalid_argument("Модуль должен быть простым числом");
     }
